// src/types/user.ts
export interface IUser {
  _id: string;
  cid: string;
  email: string;
  firstName: string;
  lastName: string;
  phone: string;
  profileImgUrl: string;
  address: string;
  district: string;
  position: string;
  birthDay: string;
  seat: string;
  createdAt: Date;
  updatedAt: Date;
  __v: number;
}